# fix: handle gcf, verify already provided request.body