--- a/src/types.ts
+++ b/src/types.ts
@@ -18,7 +18,6 @@ export type EmitterWebhookEvent<
 export type EmitterWebhookEventWithStringPayloadAndSignature = {
   id: string;
   name: EmitterWebhookEventName;
-  body?: { [key: string]: any };
   payload: string | Buffer;
   signature: string;
 };