--- a/src/verify-and-receive.ts
+++ b/src/verify-and-receive.ts
@@ -39,7 +39,7 @@ export async function verifyAndReceive(
     });
   }
 
-  let payload;
+  let payload: EmitterWebhookEvent["payload"];
 
   try {
     payload =