@wolfy1339 2023-11-29
Can you make sure that this doesn't effect values that end in a zero decimal, ex: `123.10`