@Uzlopak 2023-11-29
@wolfy1339 
I could not follow, what you mean with handling numbers correctly. 

But now it is trying to avoid a serializiation and deserialization based on the input. GCF has body and rawBody, while rawBody is the Buffer. 

Let me test this on GCF and give you feedback :).