@Uzlopak 2023-11-29
@wolfy1339 

It works now properly. Can you please explain what your concern is regarding numbers?