@wolfy1339 2023-11-29
NodeJS likes to remove trailing zeroes on numbers, and thus it changes the signature of the payload