@Uzlopak 2023-11-29
@wolfy1339 

In this case, I just handle the specific case of GCF because there is no way I can ensure, that JSON.stringify and .parse will respect that.

PTAL