@Uzlopak 2023-11-29
Tested on GCF and works as expected ;)