@wolfy1339 2023-11-30
Please let other maintainers the time to review this. Can you split the performance patches from this PR and only handle GCF compatibility please