@wolfy1339 2023-12-02
We have 3 active Pars, of which 2 were split off from this one.

Can you incorporate the feedback of this PR into the other PRs?

#939 for the switch to using `Buffer` instead of `string`
#937 for the changes to fix compatibility with Google Cloud Functions.

Let's continue on in those more specific PRs